Refactor GitHub data fetching and cleanup
- Implemented `GithubService` class to encapsulate GitHub API calls, enhancing code maintainability and reducing redundancy across API routes.
- Removed direct fetch calls from `repos.ts` API route, delegating data retrieval responsibilities to the newly created service layer, streamlining data fetching logic, and improving error handling.
- Deleted `headersApiGithub.ts` as authorization headers are now centralized within the `GithubService`, reducing code fragmentation and improving security practices by centralizing API request configurations.
- Updated `test/index.tsx` to reflect changes in data structure, particularly the renaming of the `pullrequests` field to `pullRequests`, aligning with JavaScript naming conventions and ensuring consistency across the codebase.
- Introduced type annotations and interfaces for better type safety and documentation of expected data structures from GitHub API responses, aiding in future development and maintenance efforts.
- Code cleanup involved removing unused imports, variables, and outdated comments, contributing to a cleaner, more understandable codebase.
