+++ package.json
@@ -14,7 +14,6 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@radix-ui/react-icons": "^1.3.0",
-    "@types/node": "^18.15.7",
     "@types/react": "^18.0.29",
     "@types/react-dom": "^18.0.11",
     "@vercel/analytics": "^1.1.1",
@@ -45,6 +44,7 @@
     "typescript": "^5.0.2"
   },
   "devDependencies": {
+    "@types/node": "^20.14.2",
     "@types/react-window": "^1.8.8",
     "@types/three": "^0.149.0",
     "@types/webpack-env": "^1.18.0",
